remove joint_matrix_bf16_fill_k_cache_runtime_dim_impl.hpp
